[Feed] Rename polluted histogram and fix histogram size
ContentSuggestions.Feed.TokenCompleted.ContentCount.Synthetic was
supposed to be recorded with a max value of 50 (kMaxContentCount),
but due to a mistake, it's recorded with a max size of 10.
This makes the data mostly useless, as usually values of ~35 are
recorded.

This CL change the max value to 50, and make a new histogram name.

Bug: 943838
Change-Id: I7f84bfcaec7741867f80a6d0692cdbceb5f6f786
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1542056
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Gang Wu <gangwu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#644934}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9f5cd3abf3cc3a4248ad075b2ed8584d55664588