Update Download UI according to whether we are requesting AP verdicts
This CL creates a set of strings for the download surfaces when the
user is enrolled in the Advanced Protection program. As we roll out
additional download protections for Advanced Protection users, this
will be useful for explaining the extra warnings.

It also slightly modifies the logic for when we request AP verdicts,
to make this easier to test. There are now two feature flags, one of
which is controlled by chrome://flags, and unconditionally requests
AP verdicts. The second will be used for a Finch rollout, and only
requests AP verdicts if the user is enrolled in AP, as expected.

Rendered examples: https://screenshot.googleplex.com/v3QYS1YzSsC
https://screenshot.googleplex.com/J3rWi0eK38o

Bug: 917190
Change-Id: I2a3668f914d5e04cf1f95c0e5bc5f396c569760c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1544801
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#647517}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f379e599d129e944dff590ebf53681525fc688b7